Add a new libclcore_g library which contains debug info.
Add a new libclcore_g.bc which is functionally equivalent to
libclcore.bc but is compiled with debug information and -O0.
To achieve this some allocation functions had to be written in C,
 because they were previously available in LLVM IR without debug
information.
The RenderScript cts has been run to verify these changes, both as it
is and forcing it to link with the new libclcore_g library. Both runs
show no regressions.

Change-Id: Iffc3c54fbf69b9a43703d3c2776101ef881deffd
Signed-off-by: Verena Beckham <verena@codeplay.com>