Rename memcpy to local_memcpy in rs_allocations.c.
This is so it doesn't get linked with instead of the bionic version.
This fixes a crash when compiling with -O0, where the memcpy version
in rs_allocations.c was hanging around (not being inlined) and was
linked with in unrelated code, which led to a crash because the ABI
was not the expected.

Change-Id: I0c699bf3e0ed7b353d0e2108c52fcb19660eecc5
Signed-off-by: Verena Beckham <verena@codeplay.com>