Fix NDK linking
libutils should not be used with LOCAL_SDK_VERSION, since it builds
against the platform, not the NDK, and may not be portable. It doesn't
appear to be necessary, so just remove it.

Linking against platform-built libraries in ndk-built libraries is
now a warning, and will be an error shortly.

Change-Id: Id330692098d281af6a841abf19043a281309df55