Make unique, standalone tests for all JIT expressions.
Rather than manually write the 1000s of individual testscases,
we automatically generate them and their variants to add to the
testclass.
This is done from a list of expressions that are all tested in the same
way, so a metaclass is a natural way to implement this.

Any failure in a JIT expression won't immediately abort the
`test_call_api_funs` testsuite as before, but will test all expressions
independently, and report the failures individually.

Change-Id: I10dd10398e83c45f50d9e3bbad41ad47ad84d62e
Signed-off-by: Luke Drummond <luke.drummond@codeplay.com>