[testsuite] Fix an egregious failure to report failure.
After I added the `--app-type` argument allowing a test file to be
runnable with multiple app types, I failed to realise that the running
total of pass/fails is actually implemented as a dictionary keyed by
test file name. Thus, if `test_something.py` fails for app type 'java',
but then succeeds for the JNI version, the final exit code of `run_tests.py`
will be successful as the success result for the second running of
`test_something.py` will overwrite the first as the dictionary key will
be identical

This patch fixes that issue, asserts that all new results are not yet
recorded, and fixes the running total of tests which clouded the issue.

Change-Id: If666cead11ad934aa324e06d46f1c733ccad6f1e
Signed-off-by: Luke Drummond <luke.drummond@codeplay.com>