Localize SPIR-V ID in generated SPIR-V wrapper functions
The RS kernel wrapper generated in RSoV ReflectionPass was using the same
set of SPIR-V IDs in the boilerplate code. So, if there were more than one
kernel in the same module, their boilerplate codes would conflict. This
change fixes that problem by prepending kernel names in the generated
wrapper code in each SPIR-V ID introduced.

Wrapper-function-internal names are localized, whereas types related to
kernel signature and arguments are not localized in this CL.

This is a prerequisite of multi-kernel support.

A testcase that has hardcoded expectation of boilerplate code ID string
has been updated as well.

Test: RSoV LIT Tests, RSoVTest on angler
Change-Id: Ia51b852c511adab444ec100f88d8818e0764b45f