Make sure RSoV is off by default unless flag is set
Bug: http://b/33569705

The field in the context may be left uninitialized and select RSoV
by mistake. Fixed it by always correctly initializing it to false
in the Context constructor.

Also made a fix to correctly return error on context creation and
intialization so that a client can handle a RS context creation
failure.

Test: mm
Test: boot and tested UI
Test: fast tap on build number
Test: RSTest

Change-Id: Ib40d6715004f5379521972653e8fe9011ffe771c