Add .rs.info versionInfo parser to the cpu_ref driver
This is used by lldb to check for a versioning mismatch between the
frontent and backend, but is not used by libRS. However, libRS should
not fail in its presence, so this CL enables the runtime to skip this
section if it is present.

This field is not checked for by the compatibility lib.

Test: aosp_x86_64-eng (emulator)
 - libbcc/tests
 - RsTest (32-bit, 64-bit)
 - cts -m RenderscriptTest
 - rs/lldb (required various other pending CLs)

Bug: 23535482

Change-Id: I3056dd08e7e14ff1094ead60756fd568131bcf4d
Signed-off-by: Luke Drummond <luke.drummond@codeplay.com>
(cherry picked from commit 341f30630d310f52297cf2d9867618cb7a1e8969)