Refactor KernelSignature
This is refactor of the KernelSignature class, which used to assume there
was only one argument for any given kernel. Now a vector of strings is
used to hold the kernel argument type names in KernelSignature instead of
just a string. Also moved KernelSignature methods into a dedicated file.

This is a prerequisite of supporting kernels with multiple input
allocations as arguments.

Test: RSoV lit tests and RSoVTest
Change-Id: I403e90fa4d13eda34141c30ab003fc6e6de7593a