Fix build breakage on SDK(Mac) by avoiding an include
There's a problem with including <string.h> in rs_cl.c.
For some reason the compiler still has the system headers in the
include path, and string.h isn't any of the include paths on the command line.

For now, avoid the problem altogether by using the old way of getting bits.

Test: Compiled the module.

Change-Id: I52314bd2aa41409796403e57864a3a288fbb767d