Fix failing lldb test test_language_subcmds_no_debug.py(cpp).
This test was broken by
https://android-review.googlesource.com/#/c/322609/3/tests/lldb/cpp/WaitAttach/Android.mk

Before that change, we compiled a script twice, once with -g and once
without; after that change, we compile with -g both times.

This fix goes back to compiling once with -g and once without.

Test: aosp_arm64-eng on emulator, run rs lldb tests
test_language_subcmds_no_debug.py and test_language_subcmds.py (which
share the script in question).

Change-Id: I94b6893c8e421684e12121ef8a0be14b46893d05