Fix failing lldb test test_allocation_list.py(java).
Some (but not all) of the java tests of 'language renderscript allocation list'
show an extra initial allocation.  In my current testing, some of the
tests that used to show an extra initial allocation no longer do.
Therefore, I have modified those tests to no longer expect an extra
initial allocation.

What might be the reason for the extra initial allocation?  In the
output of 'language renderscript allocation list', each allocation is
shown with an index that indicates its creation order, and the
expected indices don't change (either between the java and non-java
versions of the test, or in the fix I'm making to get the java
versions to pass), so the extra allocation doesn't seem to be some
kind of temporary allocation that is no longer being created.  Perhaps
it is being garbage-collected away now, while that wasn't happening
before?  I have not investigated further.

Test: aosp_arm64-eng on emulator, run rs lldb test test_allocation_list.py.
Change-Id: I0601660d8abc46e9075eaafbeb305154202a9bc0