Reorganized metadata related code in rs2spirv
Bug: 30964317

Added a Context class, which is a singleton that is shared across
all LLVM and Spirit compiler passes for rs2spirv. Inside Context,
there is a bcinfo::MetadataExtractor, as well as data structures
for quick lookup of information about foreach kernels and exported
variables.

Cleaned up passes which used to use MetadataExtractor to use this
new Context.

Test: lit tests
Change-Id: I3c88f8e8e75b354d87c155312176544fb660526b