Fix RSTest_CompatLib and handling of RS_ERROR_FATAL_DEBUG
RSTest_CompatLib was calling rsAllocationCopy1DRange with a 2D allocation,
which triggered an error of type RS_ERROR_FATAL_DEBUG, which was not well
handled in the message handler, and resulted in test app hang.

This CL fixed the testcase to use the correct 2D version of
allocation copy range, and changed the message handler to recognize
RS_ERROR_FATAL_DEBUG as an fatal error.

Bug: 35628302
Test: RSTest_Compat on Angler
Change-Id: I8e35a08fadd29db82c236d3aa6da74de11eb068b