Fix crashing CPU driver runtime on null input allocation
Part of the CPU driver runtime assumed non-null input allocation and
blindly dereferenced the handle of the input allocation. This caused
crashes when running RsBalls, in which a null input allocation was given
to rsForEach(). This CL fixes the crashing problem by guarding those
codes dereferencing the input allocation with null pointer checks.

Bug: 37754476
Test: CTS with and without debug.rs.debug=1, RsBalls

Change-Id: I8bec116fa8b846c6801353234e24b07d7a0fbbf4