Add a way to specify the min build tools version for unit test
Before RSUnitTests assumed that the build tools version of Slang and the
platform API version were the same. This causes tests for Slang
regression fixes to either fail on forward compatibility tests or not
run on backward compatibility tests based on what API version it is
associated with in RSUnitTests.py.

BUG: 64115371
Test: Run RSTestForward25 and check that reflection3264 doesn't appear
in logs/classes.dex

Change-Id: I4760655777e6a328cabc170c0d9c6b408f58122c