libRS is LLNDK
CTS JniStaticTest tries to load every *.so under /system/lib and
compares the actual result with the expected result.
Dlopening /system/lib/vndk-sp/libRS.so should fail since the vndk-sp
libs are not for apps. However, the dlopen is successful because
libRS.so is in /system/etc/public.libraries.txt.

In order to fix the problem, libRS is marked as LLNDK so that there is
only one instance of libRS.so in /system/lib. This possible since libRS
is already NDK and thus its ABI is guaranteed to be stable.

Bug: 64888291
Test: JniStaticTest#test_linker_namespaces successful
Change-Id: I42f8c8883215b4c852aa5792048a2fa4fb883b76