frameworks/rs: Set LOCAL_SDK_VERSION where possible.
This change sets LOCAL_SDK_VERSION for all packages where
this is possible without breaking the build, and
LOCAL_PRIVATE_PLATFORM_APIS := true otherwise.

Setting one of these two will be made required soon, and this
is a change in preparation for that. Not setting LOCAL_SDK_VERSION
makes the app implicitly depend on the bootclasspath, which is
often not required. This change effectively makes depending on
private apis opt-in rather than opt-out.

Tests are also migrated to use stubs instead of the libraries that
contain private APIs.

Test: make relevant packages
Bug: 73535841
Merged-In: I04e2caab2614c2b7f6d5560c25b257f03cb2a81c
Change-Id: Ic82b4f92837a5ce3a06bb1c7653ce0b062e139e3